Migrate subject_settings and gpa_system_settings modules to KMP (Phase 9b)
Converts subject_settings_data, subject_settings_presentation,
gpa_system_settings_data, and gpa_system_settings_presentaion together,
since the _presentation modules depend on their _data siblings' Koin
wiring being multiplatform-compatible first.

New shared infrastructure - core/data/local/datastore/AppDataStoreFactory.kt:
DataStore's Android-only Context.dataStore delegate (backed by a
java.io.InputStream/OutputStream Serializer<T>) has no Compose
Multiplatform equivalent. The multiplatform path is Okio-backed:
OkioSerializer<T> (readFrom(BufferedSource)/writeTo(T, BufferedSink)
instead of raw JVM streams) passed to OkioStorage, fed into
DataStoreFactory.create(). Added a shared createDataStore(fileName,
serializer) in :core, mirroring the existing getDatabaseBuilder() pattern
exactly: an expect fun dataStoreFilePath(fileName): okio.Path with no
platform handle in its signature, Android's actual reaching into Koin's
GlobalContext for the Application (GlobalContext.get().get<Application>()
.filesDir.resolve(fileName).toOkioPath()) rather than threading a Context
through the DI graph. Verified the exact OkioSerializer/OkioStorage API
surface against AndroidX source directly (readFrom/writeTo signatures,
OkioStorage's constructor params) rather than assuming.

GPASerializer.kt/SubjectSettingsSerializer.kt: rewritten from
Serializer<T> to OkioSerializer<T> - source.readUtf8()/sink.writeUtf8(...)
replacing input.readBytes().decodeToString()/output.write(...), same
kotlinx.serialization Json calls otherwise.

GPADatastore.kt/SubjectSettingsDataSource.kt: dropped their Context
constructor parameter entirely (now call createDataStore(fileName,
serializer) directly) - Koin wiring updated from
single { GPADatastore(get()) } to single { GPADatastore() }, matching
DatabaseModule.kt's existing no-Context-in-DI-graph convention.

core/build.gradle.kts: added the new androidx-datastore-core-okio and
okio catalog entries as api (not implementation) since OkioSerializer/
createDataStore are part of :core's public surface that the *_data
modules implement directly.

*_presentation modules: same koinViewModel Android->multiplatform swap as
Phase 9a (org.koin.androidx.compose.koinViewModel ->
org.koin.compose.viewmodel.koinViewModel); no other Android coupling found.

Note on verification: unverified in this sandbox - no Android SDK, Gradle
9.3.1 download blocked by network policy (only 8.14.3 available). The
DataStore rewrite is the highest-risk part of this PR - needs a real build
plus a manual check that existing on-disk GPA/SubjectSettings preference
files are still readable after this change (same JSON format, same file
path resolution via Application.filesDir, so should be compatible, but
this needs verifying against a real device with existing app data, not
just a fresh install).
