+++ core/src/androidMain/kotlin/com/hussienFahmy/core/data/local/datastore/AppDataStoreFactory.android.kt
@@ -1,0 +1,11 @@
+package com.hussienfahmy.core.data.local.datastore
+
+import android.app.Application
+import okio.Path
+import okio.Path.Companion.toOkioPath
+import org.koin.core.context.GlobalContext
+
+actual fun dataStoreFilePath(fileName: String): Path {
+    val appContext = GlobalContext.get().get<Application>()
+    return appContext.filesDir.resolve(fileName).toOkioPath()
+}
