+++ gpa_system_settings/gpa_system_settings_data/build.gradle.kts
@@ -1,11 +1,14 @@
 plugins {
-    alias(libs.plugins.base.module)
+    alias(libs.plugins.base.kmp.module)
 }
 
-dependencies {
-    implementation(project(":core"))
-    implementation(project(":gpa_system_settings:gpa_system_settings_domain"))
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            implementation(project(":core"))
+            implementation(project(":gpa_system_settings:gpa_system_settings_domain"))
 
-    implementation(libs.androidx.datastore.core)
-    implementation(libs.kotlinx.serialization.json)
+            implementation(libs.kotlinx.serialization.json)
+        }
+    }
 }
