+++ gpa_system_settings/gpa_system_settings_data/src/androidMain/AndroidManifest.xml
