+++ gpa_system_settings/gpa_system_settings_data/src/commonMain/kotlin/com/hussienfahmy/gpa_system_settings_data/datastore/GPADatastore.kt
@@ -1,0 +1,26 @@
+package com.hussienfahmy.gpa_system_settings_data.datastore
+
+import com.hussienfahmy.core.data.local.datastore.createDataStore
+import com.hussienfahmy.gpa_system_settings_data.model.GPA
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
+
+internal class GPADatastore {
+    private val dataSource = createDataStore("gpa_system", GPASerializer)
+
+    fun observeGPASystem(): Flow<GPA> {
+        return dataSource.data
+    }
+
+    suspend fun getGPASystem(): GPA {
+        return dataSource.data.first()
+    }
+
+    suspend fun updateGPASystem(system: GPA.System) {
+        dataSource.updateData { settings ->
+            settings.copy(
+                system = system
+            )
+        }
+    }
+}
