+++ gpa_system_settings/gpa_system_settings_presentaion/src/androidMain/AndroidManifest.xml
