+++ gradle/libs.versions.toml
@@ -24,6 +24,7 @@ androidxActivityCompose = "1.12.4"
 room = "2.8.4"
 dataStore = "1.2.0"
 preferencesDataStore = "1.2.0"
+okio = "3.17.0"
 constraintLayoutCompose = "1.1.1"
 work = "2.11.1"
 junitVersion = "1.3.0"
@@ -112,6 +113,11 @@ androidx-room-compiler = { group = "androidx.room", name = "room-compiler", vers
 # DataStore
 androidx-datastore-core = { group = "androidx.datastore", name = "datastore", version.ref = "dataStore" }
 androidx-datastore-preferences = { group = "androidx.datastore", name = "datastore-preferences", version.ref = "preferencesDataStore" }
+# Multiplatform DataStore - Okio-backed Storage<T>, replaces the Android-only Context.dataStore
+# delegate (java.io.InputStream/OutputStream-based Serializer<T>) used by GPADatastore/
+# SubjectSettingsDataSource before their KMP migration.
+androidx-datastore-core-okio = { group = "androidx.datastore", name = "datastore-core-okio", version.ref = "dataStore" }
+okio = { group = "com.squareup.okio", name = "okio", version.ref = "okio" }
 
 # Work Manager
 androidx-work-runtime-ktx = { group = "androidx.work", name = "work-runtime-ktx", version.ref = "work" }
