+++ subject_settings/subject_settings_data/src/androidMain/AndroidManifest.xml
