+++ subject_settings/subject_settings_data/src/commonMain/kotlin/com/hussienfahmy/subject_settings_data/di/Module.kt
@@ -1,0 +1,12 @@
+package com.hussienfahmy.subject_settings_data.di
+
+import com.hussienfahmy.core.domain.subject_settings.repository.SubjectSettingsRepository
+import com.hussienfahmy.subject_settings_data.datastore.SubjectSettingsDataSource
+import com.hussienfahmy.subject_settings_data.repository.SubjectSettingsRepositoryImpl
+import org.koin.dsl.bind
+import org.koin.dsl.module
+
+val subjectSettingsDataModule = module {
+    single { SubjectSettingsDataSource() }
+    single { SubjectSettingsRepositoryImpl(get()) }.bind<SubjectSettingsRepository>()
+}
