+++ subject_settings/subject_settings_data/src/commonMain/kotlin/com/hussienfahmy/subject_settings_data/model/SubjectSettings.kt
