+++ subject_settings/subject_settings_presentation/src/androidMain/AndroidManifest.xml
