+++ subject_settings/subject_settings_presentation/src/commonMain/kotlin/com/hussienfahmy/subject_settings_presentation/di/Module.kt
